Enable configuration of custom per-project environment variables
The current configuration files does not work bacause:

- The global .dockcross is run before the container
- The per-project .dockcross is not sourced, only executed

I didn't find a way to export environment variables using gosu, that's
why creating a new .dockcross-env configuration only for environment
variables.

This seemed to be the simples approach, with the downside that as it is
currently implemented, there is no validation of what's inside the
.dockcross-env file, which could be a regular shell script with more
things than environment variables.

Alternatives - of future enhancements - could be:

- Parse .dockcross-env to check for only lines that export variables
- Use for instance a JSON instead, parsing it and then exporting the
variables from the key-value pairs.