Enable configuration of custom per-project environment variables
The current configuration file does not work bacause:

- The global .dockcross is run before the container
- The per-project .dockcross is not sourced, only executed

I didn't find a way to export environment variables using gosu. The
alternative is to log all variables exported in the .dockcross script
and re-export them after.

An alternative would be to parse the .dockcross script, which would
avoid implicitly exported variables to be re-exported.